backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pillow==11.3.0
boto3==1.40.58
django-storages==1.14.6
paypalrestsdk==1.13.3
requests==2.32.5
Pillow==11.3.0
